Add more test stubs for frontend.
